--- a/fe/fe-core/src/main/java/com/starrocks/planner/RuntimeFilterDescription.java
+++ b/fe/fe-core/src/main/java/com/starrocks/planner/RuntimeFilterDescription.java
@@ -17,6 +17,7 @@ package com.starrocks.planner;
 import com.google.common.collect.Lists;
 import com.google.common.collect.Maps;
 import com.starrocks.analysis.Expr;
+import com.starrocks.analysis.SortInfo;
 import com.starrocks.qe.SessionVariable;
 import com.starrocks.thrift.TNetworkAddress;
 import com.starrocks.thrift.TRuntimeFilterBuildJoinMode;
@@ -73,6 +74,8 @@ public class RuntimeFilterDescription {
     // join's equal conjuncts size > 1.
     private final Map<Integer, List<Expr>> nodeIdToParitionByExprs = Maps.newHashMap();
 
+    private SortInfo sortInfo;
+
     public RuntimeFilterDescription(SessionVariable sv) {
         nodeIdToProbeExpr = new HashMap<>();
         mergeNodes = new ArrayList<>();
@@ -122,6 +125,14 @@ public class RuntimeFilterDescription {
         this.type = type;
     }
 
+    public SortInfo getSortInfo() {
+        return sortInfo;
+    }
+
+    public void setSortInfo(SortInfo sortInfo) {
+        this.sortInfo = sortInfo;
+    }
+
     public boolean canProbeUse(PlanNode node) {
         if (!canAcceptFilter(node)) {
             return false;
@@ -148,10 +159,24 @@ public class RuntimeFilterDescription {
 
     // return true if Node could accept the Filter
     public boolean canAcceptFilter(PlanNode node) {
-        if (RuntimeFilterType.TOPN_FILTER.equals(runtimeFilterType()) && !(node instanceof OlapScanNode)) {
+        if (RuntimeFilterType.TOPN_FILTER.equals(runtimeFilterType())) {
+            if (node instanceof ScanNode) {
+                ScanNode scanNode = (ScanNode) node;
+                return scanNode.isOlapScanNode() || scanNode.isExternalTableScanNodeWithFileSystem();
+            } else {
+                return false;
+            }
+        }
+
+        return true;
+    }
+
+    public boolean isNullLast() {
+        if (sortInfo != null) {
+            return !sortInfo.getNullsFirst().get(0);
+        } else {
             return false;
         }
-        return true;
     }
 
     public void enterExchangeNode() {