--- a/fe/fe-core/src/main/java/com/starrocks/planner/ScanNode.java
+++ b/fe/fe-core/src/main/java/com/starrocks/planner/ScanNode.java
@@ -146,4 +146,13 @@ public abstract class ScanNode extends PlanNode {
 
         return columnAccessPaths.stream().map(ColumnAccessPath::toThrift).collect(Collectors.toList());
     }
+
+    public boolean isExternalTableScanNodeWithFileSystem() {
+        return this instanceof HdfsScanNode || this instanceof IcebergScanNode || this instanceof HudiScanNode ||
+                this instanceof PaimonScanNode || this instanceof DeltaLakeScanNode;
+    }
+
+    public boolean isOlapScanNode() {
+        return this instanceof OlapScanNode;
+    }
 }