Fixes #issue
Because following two limitations when using topn rf.
1. Currently we don't support `or` predicate filtering.
2. External table column must be nullable.

So we can only build topn rf in case of null last on external table. Currently if the user can confirm that there is no null in sort column and want to use this optimization, they can
1. modify sql to "order by c1 NULLS LAST"
2. set sql_mode='SORT_NULLS_LAST';
3. set enable_external_table_topn_rf_with_null_first = true (Currently, it only supports order by column and cannot support order by expr, so this method is not supported now. )

Some test cases:
1 fe 3be 8c 16g tpch 1TB hive external table


case 1 :  `select l_partkey, l_orderkey from lineitem order by l_orderkey NULLS LAST  limit 10`
| file format | base | patched |
|--------|--------|--------|
| parquet   | 75s | 2s |
| orc   | 60s | 3s |

case 2: ` select l_partkey, l_orderkey from lineitem order by l_partkey NULLS LAST  limit 10;`

 file format | base | patched |
|--------|--------|--------|
| parquet   | 85s | 3s |
| orc   | 62s | 3.5s |

case 3 ` select l_orderkey,l_suppkey from lineitem order by l_suppkey  NULLS LAST limit 10;`

 file format | base | patched |
|--------|--------|--------|
| parquet   | 73s | 7s |
| orc   | 73s | 8.4s |

case 4 ` SELECT
    *
FROM
    (
        SELECT
            l.l_partkey, l.l_orderkey,
            p.p_name,
            p.p_brand
        FROM
            lineitem l
            LEFT JOIN part p ON l.l_partkey = p.p_partkey
    ) AS mocktable
ORDER BY
    mocktable.l_partkey NULLS LAST
LIMIT
    20;`

 file format | base | patched |
|--------|--------|--------|
| parquet   | 68s | 8s |
| orc   | 73s | 15s |

I think the effect will be better after parquet page index is supported.

## What type of PR is this:

- [ ] BugFix
- [ ] Feature
- [x] Enhancement
- [ ] Refactor
- [ ] UT
- [ ] Doc
- [ ] Tool

Does this PR entail a change in behavior?

- [ ] Yes, this PR will result in a change in behavior.
- [x] No, this PR will not result in a change in behavior.

If yes, please specify the type of change:

- [ ] Interface/UI changes: syntax, type conversion, expression evaluation, display information
- [ ] Parameter changes: default values, similar parameters but with different default values
- [ ] Policy changes: use new policy to replace old one, functionality automatically enabled
- [ ] Feature removed
- [ ] Miscellaneous: upgrade & downgrade compatibility, etc.

## Checklist:

- [ ] I have added test cases for my bug fix or my new feature
- [ ] This pr needs user documentation (for new or modified features or behaviors)
  - [ ] I have added documentation for my new feature or new function

## Bugfix cherry-pick branch check:

- [x] I have checked the version labels which the pr will be auto-backported to the target branch
  - [ ] 3.1
  - [ ] 3.0
  - [ ] 2.5
  - [ ] 2.4