@wanpengfei-git 2023-09-13
#### [FE Incremental Coverage Report]

:heart_eyes: **pass** : 13 / 15 (86.67%)




#### file detail

|   |path|covered_line|new_line|coverage|not_covered_line_detail|
|----|----|----|----|----|----|
|:large_blue_circle:|com/starrocks/planner/RuntimeFilterDescription.java|9|11|81.82%|[129, 178]|
|:large_blue_circle:|com/starrocks/planner/SortNode.java|1|1|100.00%|[]|
|:large_blue_circle:|com/starrocks/planner/ScanNode.java|2|2|100.00%|[]|
|:large_blue_circle:|com/starrocks/planner/PlanNode.java|1|1|100.00%|[]|